fix: coloca as paginas como paginas
